requirements.txt
huggingface-hub==0.27.1
neptune==1.13.0
neptune_tensorflow_keras==2.2.2
pandas
pydantic-settings==2.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dotenv>=1.0.1
tensorflow==2.18.0
yfinance==0.2.50